wechat-auto/requirements.txt
# 企业微信自动回复系统依赖

# 核心依赖
requests>=2.28.0
PyYAML>=6.0

# 文档处理（可选）
PyPDF2>=3.0.0          # PDF文件读取
python-docx>=0.8.11    # Word文档读取

# 高级功能（后续可添加）
# sentence-transformers>=2.2.0  # 文本向量化（用于语义检索）
# chromadb>=0.4.0              # 向量数据库
# openai>=1.0.0               # OpenAI API（如果需要GPT增强）
# langchain>=0.0.200          # LLM应用框架

# 开发工具
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0